devices_v2/pyproject.toml
[project]
name = "cash_system"
version = "0.1.0"
requires-python = ">=3.12,<3.13"

dependencies = [
    "anyio==4.9.0",
    "async-timeout==5.0.1",
    "certifi==2025.1.31",
    "colorlog==6.9.0",
    "elastic-transport==8.17.1",
    "elasticsearch==8.17.2",
    "exceptiongroup==1.2.2",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "idna==3.10",
    "mpmath==1.3.0",
    "pycryptodome==3.22.0",
    "pyserial==3.5",
    "pyserial-asyncio==0.6",
    "redis==5.2.1",
    "sniffio==1.3.1",
    "sympy==1.13.3",
    "typing-extensions==4.13.2",
    "urllib3==2.3.0",
    "websockets==15.0.1",
]
